[WFLY-12304]: Upgrade Apache Artemis from 2.8.1 to 2.10.1
* Updating artemis version
* Adding new dependency on jctools-core
* Adding a pure netty failover test

Jira: https://issues.jboss.org/browse/WFLY-12304

+++ messaging-activemq/src/main/java/org/wildfly/extension/messaging/activemq/AddressSettingAdd.java
@@ -119,7 +119,7 @@ class AddressSettingAdd extends AbstractAddStepHandler {
             settings.setRedistributionDelay(AddressSettingDefinition.REDISTRIBUTION_DELAY.resolveModelAttribute(context, config).asLong());
         }
         if (config.hasDefined(AddressSettingDefinition.PAGE_SIZE_BYTES.getName())) {
-            settings.setPageSizeBytes(AddressSettingDefinition.PAGE_SIZE_BYTES.resolveModelAttribute(context, config).asLong());
+            settings.setPageSizeBytes(AddressSettingDefinition.PAGE_SIZE_BYTES.resolveModelAttribute(context, config).asInt());
         }
         if (config.hasDefined(AddressSettingDefinition.PAGE_MAX_CACHE_SIZE.getName())) {
             settings.setPageCacheMaxSize(AddressSettingDefinition.PAGE_MAX_CACHE_SIZE.resolveModelAttribute(context, config).asInt());
